Got wasm instantiated in a worker, and source maps working for dist
