+++ lib/wasmboy/fetchrom.js
