+++ lib/wasmboy/wasmboy.js
@@ -4,14 +4,11 @@ import Promise from 'promise-polyfill';
 
 // WasmBoy Modules
 import { WasmBoyGraphics } from '../graphics/graphics';
-import { WasmBoyAudio } from '../audio/audio';
-import { WasmBoyController } from '../controller/controller';
 import { WasmBoyMemory } from '../memory/memory';
 
 // Other lib helpers
-import { instantiateWasm } from '../wasm/instantiate';
 import { instantiateWorkers } from '../worker/instantiate';
-import { fetchROMAsByteArray } from './rom';
+import { loadROMToWasmBoy } from './loadrom';
 import { render } from './render';
 import { update } from './update';
 
@@ -22,8 +19,6 @@ const raf = require('raf');
 class WasmBoyLibService {
   // Start the request to our wasm module
   constructor() {
-    this.wasmInstance = undefined;
-    this.wasmByteMemory = undefined;
     this.worker = undefined;
     this.canvasElement = undefined;
     this.paused = false;
@@ -114,87 +109,8 @@ class WasmBoyLibService {
 
   // Finish request for wasm module, and fetch game
   loadROM(ROM, fetchHeaders) {
-    // Getting started with wasm
-    // http://webassembly.org/getting-started/js-api/
-    this.ready = false;
-    this.loadedAndStarted = false;
-
-    const initializeTask = async () => {
-      // Get our promises
-      const initPromises = [fetchROMAsByteArray(ROM, fetchHeaders), this._instantiateWorkers(), this._instantiateWasm()];
-
-      if (!this.options.headless && WasmBoyMemory.getLoadedCartridgeMemoryState().RAM) {
-        initPromises.push(WasmBoyMemory.saveCartridgeRam());
-      }
-
-      return await Promise.all(initPromises);
-    };
-
-    const loadROMAndConfigTask = async responses => {
-      // Clear what is currently in memory, then load the cartridge memory
-      WasmBoyMemory.clearMemory();
-      WasmBoyMemory.resetState();
-
-      // TODO: Handle passing a boot rom
-      WasmBoyMemory.loadCartridgeRom(responses[0]);
-
-      // Save the game that we loaded if we need to reload the game
-      this.loadedROM = responses[0];
-
-      // Run our initialization on the core
-      this.wasmInstance.exports.config(
-        0, // TODO: Include Boot Rom
-        this.options.isGbcEnabled ? 1 : 0,
-        this.options.audioBatchProcessing ? 1 : 0,
-        this.options.graphicsBatchProcessing ? 1 : 0,
-        this.options.timersBatchProcessing ? 1 : 0,
-        this.options.graphicsDisableScanlineRendering ? 1 : 0,
-        this.options.audioAccumulateSamples ? 1 : 0,
-        this.options.tileRendering ? 1 : 0,
-        this.options.tileCaching ? 1 : 0
-      );
-    };
-
-    const loadROMTask = async () => {
-      // Pause wasmBoy
-      await this.pause();
-
-      // Initialize any needed parts of wasmboy
-      let responses = await initializeTask();
-
-      // Check if we are running headless
-      if (this.options.headless) {
-        await WasmBoyMemory.initialize(this.options.headless, this.wasmInstance, this.wasmByteMemory, this.options.saveStateCallback);
-
-        await loadROMAndConfigTask(responses);
-
-        this.ready = true;
-        if (this.options.onReady) {
-          this.options.onReady();
-        }
-      } else {
-        // Finally intialize all of our services
-        // Initialize our services
-        await Promise.all([
-          WasmBoyGraphics.initialize(this.canvasElement, this.wasmInstance, this.wasmByteMemory, this.options.updateGraphicsCallback),
-          WasmBoyAudio.initialize(this.wasmInstance, this.wasmByteMemory, this.options.updateAudioCallback),
-          WasmBoyController.initialize(this.wasmInstance),
-          WasmBoyMemory.initialize(this.options.headless, this.wasmInstance, this.wasmByteMemory, this.options.saveStateCallback)
-        ]);
-
-        await loadROMAndConfigTask(responses);
-
-        // Load the game's cartridge ram
-        await WasmBoyMemory.loadCartridgeRam();
-
-        this.ready = true;
-        if (this.options.onReady) {
-          this.options.onReady();
-        }
-      }
-    };
-
-    return loadROMTask();
+    const boundLoadROM = loadROMToWasmBoy.bind(this);
+    return boundLoadROM(ROM, fetchHeaders);
   }
 
   // function to start/resume
@@ -384,21 +300,6 @@ class WasmBoyLibService {
     };
 
     return instantiateWorkersTask();
-  }
-
-  // Wrapper around instantiateWasm() to ensure we don't already have the instance
-  _instantiateWasm() {
-    const instantiateWasmTask = async () => {
-      if (this.wasmInstance) {
-        return;
-      } else {
-        let response = await instantiateWasm();
-        this.wasmInstance = response.instance;
-        this.wasmByteMemory = response.byteMemory;
-      }
-    };
-
-    return instantiateWasmTask();
   }
 }
