+++ lib/worker/constants.js
@@ -1,0 +1,4 @@
+export const WORKER_MESSAGE_TYPE = {
+  CONNECT: 'CONNECT',
+  INSTANTIATE_WASM: 'INSTANTIATE_WASM'
+};
