+++ lib/worker/instantiate.js
@@ -13,6 +13,7 @@ import wasmboyMemoryWorkerUrl from '../../dist/memory/worker/memory.worker.js';
 
 // Import our Smart Worker Interface
 import SmartWorker from './smartworker';
+import { WORKER_MESSAGE_TYPE } from './constants';
 
 /*ROLLUP_REPLACE_NODE
 const { MessageChannel } = require('worker_threads');
@@ -62,14 +63,14 @@ export const instantiateWorkers = async () => {
       // Post our connect messages
       libWorker.postMessage(
         {
-          command: 'connect',
+          type: WORKER_MESSAGE_TYPE.CONNECT,
           workerId: childWorker.id
         },
         [messageChannel.port1]
       );
       childWorker.postMessage(
         {
-          command: 'connect',
+          type: WORKER_MESSAGE_TYPE.CONNECT,
           workerId: libWorker.id
         },
         [messageChannel.port2]
