+++ lib/worker/smartworker.js
@@ -52,6 +52,13 @@ export default class SmartWorker {
     });
   }
 
+  postMessageAndWaitForResponse(message) {
+    return new Promise(resolve => {
+      this.addMessageListener(resolve, true);
+      this.worker.postMessage(message);
+    });
+  }
+
   addMessageListener(callback, shouldListenOnce) {
     this.messageListeners.push({
       id: generateId(),
