+++ package.json
@@ -22,9 +22,9 @@
     "build": "npm run core:build && npm run lib:build",
     "deploy": "npm run lib:deploy && npm run debugger:deploy",
     "test": "npm run test:accuracy",
-    "test:accuracy": "npm run build && npx --node-arg=--experimental-worker mocha test/accuracy/accuracy-test.js --exit",
+    "test:accuracy": "npm run build && node --experimental-worker node_modules/mocha/bin/_mocha test/accuracy/accuracy-test.js --exit",
     "test:perf": "npm run test:performance",
-    "test:performance": "npm run build && npx --node-arg=--experimental-worker mocha test/performance/performance-test.js --exit",
+    "test:performance": "npm run build && node --experimental-worker node_modules/mocha/bin/_mocha test/performance/performance-test.js --exit",
     "debugger:watch": "preact watch --src debugger",
     "debugger:serve": "npm run debugger:build && preact serve",
     "debugger:build": "preact build --src debugger --no-prerender",
@@ -105,6 +105,7 @@
     "rollup-plugin-regenerator": "^0.5.4",
     "rollup-plugin-replace": "^2.1.0",
     "rollup-plugin-url": "^1.4.0",
+    "source-map-loader": "^0.2.4",
     "uglifyjs-webpack-plugin": "^1.2.3",
     "url-loader": "^1.0.1",
     "watch": "^1.0.2"
